requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
mysql-connector-python==8.2.0
PyPDF2==3.0.1
sentence-transformers==2.3.1
numpy==1.24.3
torch==2.1.2
transformers==4.35.2
google-generativeai==0.3.2
python-dotenv==1.0.0
